Revert "Fitting User-defined Quantities"
